+++ Documentation/Changelog.md
@@ -6,7 +6,9 @@
 
 Release date: not released yet
 
-- Added polyfill for `System.Security.Cryptography.CryptographicOperations` class
+- Polyfills:
+    - Added polyfill for `System.Security.Cryptography.CryptographicOperations` class
+    - Added polyfills for `Fill(Span<byte>)`, `GetBytes(Span<byte>)`, and `GetBytes(int)` methods of `System.Security.Cryptography.RandomNumberGenerator` class
 - Fixed issue with signature of `Process.WaitForExit(int, CancellationToken)` polyfill method that was not returning `bool` result as it should have been
 
 ### Gapotchenko.FX 2026.7